requirements.txt
py-clob-client>=0.34.0
python-dotenv>=1.0.0
requests>=2.31.0
websocket-client>=1.7.0
